+++ extensions/gyazo-uploader/CHANGELOG.md
@@ -1,3 +1,15 @@
 # Gyazo Uploader Changelog
 
 ## [Initial Version] - 2023-08-12
+
+## [1.0.0] - 2025-04-02
+
+# Features
+
+This plugin allows you to upload any image to Gyazo directly from your clipboard.
+
+[![Image from Gyazo](https://t.gyazo.com/teams/nota/ee411b63a69c0677eff4faea513168bd.gif)](https://nota.gyazo.com/ee411b63a69c0677eff4faea513168bd)
+
+You can copy as Markdown text from menubar. it for quickly sharing gyazo link to github PR's, issues, etc.
+
+[![Image from Gyazo](https://t.gyazo.com/teams/nota/50834647d8e86414cc45be87cf595b61.gif)](https://nota.gyazo.com/50834647d8e86414cc45be87cf595b61)